Refactor: Move the terminal code to its own module
